+++ MotivationGarden/src/model/Cow.java
@@ -9,7 +9,7 @@ public class Cow extends Animal {
 	// x and y position, imagePath (file path to image that GUI will draw)
 	
 	//String type;
-	int moveChance;
+	double moveChance;
 	
 	public Cow(int x, int y) {
 		this.positionX = x;
